requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
apptools=4.4.0=py36_2
arrow=0.12.1=py36_1002
asn1crypto=0.24.0=py36_1003
backcall=0.1.0=py_0
binaryornot=0.4.4=py_1
blas=1.0=mkl
bleach=3.0.0=py_0
bzip2=1.0.6=h470a237_2
ca-certificates=2018.11.29=ha4d7672_0
certifi=2018.11.29=py36_1000
cffi=1.11.5=py36h5e8e0c9_1
chardet=3.0.4=py36_1003
click=7.0=py_0
configobj=5.0.6=py_0
cookiecutter=1.6.0=py36_1000
cryptography=2.3.1=py36hdffb7b8_0
cryptography-vectors=2.3.1=py36_1000
curl=7.61.0=h93b3f91_2
cycler=0.10.0=py_1
dbus=1.13.0=h3a4f0e9_0
decorator=4.3.0=py_0
entrypoints=0.2.3=py36_1002
envisage=4.7.0=py_0
expat=2.2.5=hfc679d8_2
fontconfig=2.13.1=h65d0f4c_0
freetype=2.9.1=h6debe1e_4
future=0.16.0=py36_1002
gettext=0.19.8.1=h5e8e0c9_1
glib=2.55.0=h464dc38_2
gmp=6.1.2=hfc679d8_0
gst-plugins-base=1.12.5=hde13a9d_0
gstreamer=1.12.5=h61a6719_0
h5py=2.8.0=py36h470a237_0
hdf4=4.2.13=h951d187_2
hdf5=1.10.1=2
html5lib=1.0.1=py_0
icu=58.2=hfc679d8_0
idna=2.8=py36_1000
imageio=2.4.1=py36_1000
intel-openmp=2019.0=118
ipykernel=5.0.0=pyh24bf2e0_1
ipython=7.0.1=py36h24bf2e0_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.4.2=py_0
jedi=0.13.1=py36_1000
jinja2=2.10=py_1
jinja2-time=0.2.0=py_2
jpeg=9c=h470a237_1
jsoncpp=1.8.1=h2d50403_1
jsonschema=2.6.0=py36_1002
jupyter=1.0.0=py_1
jupyter_client=5.2.3=py_1
jupyter_console=6.0.0=py_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=py36_0
jupyter_core=4.4.0=py_0
jupyter_highlight_selected_word=0.2.0=py36_1000
jupyter_latex_envs=1.4.4=py36_1000
jupyter_nbextensions_configurator=0.4.1=py36_0
kiwisolver=1.0.1=py36h2d50403_2
krb5=1.14.6=0
libffi=3.2.1=hfc679d8_5
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h470a237_3
libnetcdf=4.6.1=2
libogg=1.3.2=h470a237_1
libopenblas=0.3.3=h5a2b251_3
libpng=1.6.35=ha92aebf_2
libsodium=1.0.16=h470a237_1
libssh2=1.8.0=h5b517e9_2
libstdcxx-ng=7.2.0=hdf63c60_3
libtheora=1.1.1=h470a237_1
libtiff=4.0.9=he6b73bb_2
libuuid=2.32.1=h470a237_2
libvorbis=1.3.5=0
libxcb=1.13=h470a237_2
libxml2=2.9.8=h422b904_5
libxslt=1.1.32=h88dbc4e_2
lxml=4.3.0=py36hc9114bc_0
lz4-c=1.8.1.2=0
markupsafe=1.0=py36h470a237_1
matplotlib=3.0.2=py36h8a2030e_1
matplotlib-base=3.0.2=py36h20b835b_1
mayavi=4.6.2=py36h18b3941_1
missingno=0.4.1=py_0
mistune=0.8.3=py36h470a237_2
mkl=2018.0.3=1
mkl_fft=1.0.10=py36_0
mkl_random=1.0.2=py36_0
nbconvert=5.3.1=py_1
nbformat=4.4.0=py_1
ncurses=6.1=hfc679d8_1
nibabel=2.3.2=pyh24bf2e0_0
nilearn=0.5.0=pyh24bf2e0_0
notebook=5.7.0=py36_1000
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py_0
openblas=0.3.3=ha44fe06_1
openssl=1.0.2p=h14c3975_1002
pandas=0.23.4=py36hf8a1672_0
pandoc=2.3.1=0
pandocfilters=1.4.2=py_1
parso=0.3.1=py_0
patsy=0.5.0=py_1
pcre=8.41=hfc679d8_3
pexpect=4.6.0=py36_1000
pickleshare=0.7.5=py36_1000
pillow=5.3.0=py36hc736899_0
pip=18.1=py36_1000
poyo=0.4.2=py_0
prometheus_client=0.4.0=py_0
prompt_toolkit=2.0.5=py_0
pthread-stubs=0.4=h470a237_1
ptyprocess=0.6.0=py36_1000
pycparser=2.19=py_0
pydicom=1.2.0=py_0
pyface=6.0.0=py_1
pygments=2.2.0=py_1
pyopenssl=18.0.0=py36_1000
pyparsing=2.2.2=py_0
pyqt=5.6.0=py36h8210e8a_7
pysocks=1.6.8=py36_1002
pysurfer=0.9.0=pyh24bf2e0_0
python=3.6.6=h5001a0f_0
python-dateutil=2.7.3=py_0
pytz=2018.5=py_0
pyyaml=3.13=py36h470a237_1
pyzmq=17.1.2=py36hae99301_0
qt=5.6.2=hf70d934_9
qtconsole=4.4.1=py36_1
readline=7.0=haf1bffa_1
requests=2.21.0=py36_1000
scikit-learn=0.19.1=py36hedc7406_0
scipy=1.1.0=py36hc49cb51_0
seaborn=0.9.0=py_0
send2trash=1.5.0=py_0
setuptools=40.4.3=py36_0
simplegeneric=0.8.1=py_1
sip=4.18.1=py36hfc679d8_0
six=1.11.0=py36_1001
sqlite=3.25.2=hb1c47c0_0
statsmodels=0.9.0=py36_0
tbb=2019.1=h2d50403_0
terminado=0.8.1=py36_1001
testpath=0.4.2=py36_1000
tk=8.6.8=ha92aebf_0
tornado=5.1.1=py36h470a237_0
tqdm=4.29.0=py_0
traitlets=4.3.2=py36_1000
traits=4.6.0=py36h470a237_2
traitsui=6.0.0=py_1
urllib3=1.24.1=py36_1000
vtk=8.1.1=py36h8649a91_204
watermark=1.8.0=py_0
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
wheel=0.32.1=py36_0
whichcraft=0.5.2=py_1
widgetsnbextension=3.4.2=py36_1000
xorg-kbproto=1.0.7=h470a237_2
xorg-libice=1.0.9=h470a237_4
xorg-libsm=1.2.2=h8c8a85c_6
xorg-libx11=1.6.6=h470a237_0
xorg-libxau=1.0.8=h470a237_6
xorg-libxdmcp=1.1.2=h470a237_7
xorg-libxt=1.1.5=h470a237_2
xorg-xproto=7.0.31=h470a237_7
xz=5.2.4=h470a237_1
yaml=0.1.7=h470a237_1
yapf=0.24.0=py_0
zeromq=4.2.5=hfc679d8_6
zlib=1.2.11=h470a237_3
